I broke it

+++ src/analyzer.c
@@ -1,21 +1,20 @@
 #include "cutter.h"
 
-// based on htop source
 static float compute_for_core(proc_stat_core_info_t* core_info)
 {
     core_info->user = core_info->user - core_info->guest;
-    core_info->nice     = core_info->nice - core_info->guest_nice;
+    core_info->nice = core_info->nice - core_info->guest_nice;
 
-    int idlealltime   = core_info->idle + core_info->iowait;
-    int systemalltime = core_info->system + core_info->irq + core_info->softirq;
-    int virtalltime   = core_info->guest + core_info->guest_nice;
+    long long idlealltime   = core_info->idle + core_info->iowait;
+    long long systemalltime = core_info->system + core_info->irq + core_info->softirq;
+    long long virtalltime   = core_info->guest + core_info->guest_nice;
 
-    int total = core_info->user + core_info->nice + systemalltime + idlealltime + core_info->steal + virtalltime;
+    long long total = core_info->user + core_info->nice + systemalltime + idlealltime + core_info->steal + virtalltime;
 
-    float busy_time = (float)total - (float)idlealltime;
-    float busy_perc = busy_time / (float)total;
+    double busy_time = (double)total - (double)idlealltime;
+    double busy_perc = busy_time / (double)total;
 
-    return busy_perc * 100.f;
+    return (float)(busy_perc * 100.f);
 }
 
 status_t analyzer_compute_usage(proc_stat_info_t* ps_info, computed_info_t* c_info)
@@ -42,10 +41,12 @@ static void* analyzer_job(void* data)
         sem_wait(&syncs->c_info_empty_semaphore);
         pthread_mutex_lock(&syncs->ps_info_mutex);
         pthread_mutex_lock(&syncs->c_info_mutex);
-            analyzer_compute_usage(&buffers->ps_infos[buffers->num_ps_infos - 1], &buffers->c_infos[buffers->num_c_infos]);
-            buffers->num_ps_infos--;
-            buffers->num_c_infos++;
-            // logger_log(LOG_INFO, "analyzer_job running!\n");
+
+        analyzer_compute_usage(&buffers->ps_infos[buffers->num_ps_infos - 1], &buffers->c_infos[buffers->num_c_infos]);
+        buffers->num_ps_infos--;
+        buffers->num_c_infos++;
+
+        // logger_log(LOG_INFO, "analyzer_job running!\n");
         pthread_mutex_unlock(&syncs->ps_info_mutex);
         pthread_mutex_unlock(&syncs->c_info_mutex);
         sem_post(&syncs->ps_info_empty_semaphore);
